refactor: remove unneeded inline functions
