--- kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/inline/runAnyAllNoneTests.kt
@@ -1,129 +1,0 @@
-package xyz.lbres.kotlinutils.set.multiset.inline
-
-import xyz.lbres.kotlinutils.int.ext.isNegative
-import xyz.lbres.kotlinutils.set.multiset.* // ktlint-disable no-wildcard-imports no-unused-imports
-import kotlin.test.assertFalse
-import kotlin.test.assertTrue
-
-internal fun runAnyTests() {
-    val listSet = multiSetOf(listOf(), listOf(), listOf(1, 2, 3, 4), listOf(1, 2, 3), listOf(5))
-    val intSet = multiSetOf(1, 2, 3, 4, 4, 4)
-
-    // empty
-    assertFalse { emptyMultiSet<Int>().any { true } }
-
-    // none
-    assertFalse { intSet.any(Int::isNegative) }
-    assertFalse { intSet.any { intSet.getCountOf(it) == 2 } }
-
-    assertFalse { listSet.any { it.contains(6) } }
-
-    // all
-    assertTrue { intSet.any { true } }
-    assertTrue { intSet.any { it in 1..7 } }
-
-    assertTrue { listSet.any { it.size < 5 } }
-
-    // some
-    assertTrue { intSet.any { it == 3 } }
-    assertTrue { intSet.any { intSet.getCountOf(it) > 1 } }
-
-    assertTrue { listSet.any { it.isEmpty() } }
-
-    // modified
-    var no4 = true
-    assertTrue {
-        intSet.any {
-            when {
-                it == 4 && no4 -> {
-                    no4 = false
-                    true
-                }
-                it == 4 -> false
-                else -> true
-            }
-        }
-    }
-}
-
-internal fun runAllTests() {
-    val listSet = multiSetOf(listOf(), listOf(), listOf(1, 2, 3, 4), listOf(1, 2, 3), listOf(5))
-    val intSet = multiSetOf(1, 2, 3, 4, 4, 4)
-
-    // empty
-    assertTrue { emptyMultiSet<Int>().all { true } }
-
-    // none
-    assertFalse { intSet.all(Int::isNegative) }
-    assertFalse { intSet.all { intSet.getCountOf(it) == 2 } }
-
-    assertFalse { listSet.all { it.contains(6) } }
-
-    // all
-    assertTrue { intSet.all { true } }
-    assertTrue { intSet.all { it in 1..7 } }
-
-    assertTrue { listSet.all { it.size < 5 } }
-
-    // some
-    assertFalse { intSet.all { it == 3 } }
-    assertFalse { intSet.all { intSet.getCountOf(it) > 1 } }
-
-    assertFalse { listSet.all { it.isEmpty() } }
-
-    // modified
-    var no4 = true
-    assertFalse {
-        intSet.all {
-            when {
-                it == 4 && no4 -> {
-                    no4 = false
-                    true
-                }
-                it == 4 -> false
-                else -> true
-            }
-        }
-    }
-}
-
-internal fun runNoneTests() {
-    val listSet = multiSetOf(listOf(), listOf(), listOf(1, 2, 3, 4), listOf(1, 2, 3), listOf(5))
-    val intSet = multiSetOf(1, 2, 3, 4, 4, 4)
-
-    // empty
-    assertTrue { emptyMultiSet<Int>().none { true } }
-
-    // none
-    assertTrue { intSet.none(Int::isNegative) }
-    assertTrue { intSet.none { intSet.getCountOf(it) == 2 } }
-
-    assertTrue { listSet.none { it.contains(6) } }
-
-    // none
-    assertFalse { intSet.none { true } }
-    assertFalse { intSet.none { it in 1..7 } }
-
-    assertFalse { listSet.none { it.size < 5 } }
-
-    // some
-    assertFalse { intSet.none { it == 3 } }
-    assertFalse { intSet.none { intSet.getCountOf(it) > 1 } }
-
-    assertFalse { listSet.none { it.isEmpty() } }
-
-    // modified
-    var previous4 = false
-    assertFalse {
-        intSet.none {
-            when {
-                it == 4 && !previous4 -> {
-                    previous4 = true
-                    false
-                }
-                it == 4 -> true
-                else -> false
-            }
-        }
-    }
-}
