+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/inline/runFilterTests.kt
@@ -9,94 +9,6 @@ private val e2 = ArithmeticException()
 private val e3 = ClassCastException("Cannot cast Int to List")
 private val e4 = ClassCastException("other message")
 
-internal fun runFilterTests() {
-    var intSet = emptyMultiSet<Int>()
-    var intExpected = emptyList<Int>()
-    assertEquals(intExpected, intSet.filter { true })
-
-    intSet = multiSetOf(1, 1, 1, 2, 3, 4, 5, 5, 6, -1, 0, 0)
-    intExpected = listOf(1, 1, 1, 2, 3, 4, 5, 5, 6, -1, 0, 0)
-    assertEquals(intExpected.sorted(), intSet.filter { true }.sorted())
-
-    intExpected = emptyList()
-    assertEquals(intExpected, intSet.filter { false })
-
-    intExpected = listOf(1, 1, 1, 5, 5, 0, 0)
-    assertEquals(intExpected.sorted(), intSet.filter { intSet.getCountOf(it) > 1 }.sorted())
-
-    intExpected = listOf(1, 1, 1, 3, -1, 6)
-    val testInts = setOf(-1, 1, 11, 12, 10, 3, 6, -2)
-    assertEquals(intExpected.sorted(), intSet.filter { it in testInts }.sorted())
-
-    val stringSet = multiSetOf("abc", "abc", "hello", "world", "goodbye", "world", "hi", "world")
-    val stringExpected = listOf("abc", "abc", "hello", "world", "world", "world")
-    assertEquals(stringExpected.sorted(), stringSet.filter { it.length in 3..5 }.sorted())
-
-    val errorSet: MultiSet<Exception> = multiSetOf(e1, e1, e2, e3, e4, e4)
-    val errorExpected: List<Exception> = listOf(e3, e4, e4)
-    assertEquals(errorExpected, errorSet.filter { it is ClassCastException }.sortedBy { it.message!! })
-
-    // modified
-    intSet = multiSetOf(1, 1, 2, 14, 14)
-    intExpected = listOf(1, 2, 14, 14)
-    var previousOdd = false
-    val intActual = intSet.filter {
-        when {
-            it % 2 == 0 -> true
-            previousOdd -> false
-            else -> {
-                previousOdd = true
-                true
-            }
-        }
-    }.sorted()
-    assertEquals(intExpected, intActual)
-}
-
-internal fun runFilterNotTests() {
-    var intSet = emptyMultiSet<Int>()
-    var intExpected = emptyList<Int>()
-    assertEquals(intExpected, intSet.filterNot { true })
-
-    intSet = multiSetOf(1, 1, 1, 2, 3, 4, 5, 5, 6, -1, 0, 0)
-    intExpected = emptyList()
-    assertEquals(intExpected, intSet.filterNot { true })
-
-    intExpected = listOf(1, 1, 1, 2, 3, 4, 5, 5, 6, -1, 0, 0)
-    assertEquals(intExpected.sorted(), intSet.filterNot { false }.sorted())
-
-    intExpected = listOf(2, 3, 4, 6, -1)
-    assertEquals(intExpected.sorted(), intSet.filterNot { intSet.getCountOf(it) > 1 }.sorted())
-
-    intExpected = listOf(2, 4, 5, 5, 0, 0)
-    val testInts = setOf(-1, 1, 11, 12, 10, 3, 6, -2)
-    assertEquals(intExpected.sorted(), intSet.filterNot { it in testInts }.sorted())
-
-    val stringSet = multiSetOf("abc", "abc", "hello", "world", "goodbye", "world", "hi", "world")
-    val stringExpected = listOf("goodbye", "hi")
-    assertEquals(stringExpected, stringSet.filterNot { it.length in 3..5 }.sorted())
-
-    val errorSet: MultiSet<Exception> = multiSetOf(e1, e1, e2, e3, e4, e4)
-    val errorExpected: List<Exception> = listOf(e1, e1, e2)
-    assertEquals(errorExpected, errorSet.filterNot { it is ClassCastException }.sortedBy { it.message ?: "ZYX" })
-
-    // modified
-    intSet = multiSetOf(1, 1, 3, 2, 14, 14)
-    val intOptions = listOf(listOf(1, 1), listOf(1, 3))
-    var previousOdd = false
-    val intActual = intSet.filterNot {
-        when {
-            it % 2 == 0 -> true
-            previousOdd -> false
-            else -> {
-                previousOdd = true
-                true
-            }
-        }
-    }.sorted()
-    assertEqualsAnyOf(intOptions, intActual)
-}
-
 internal fun runFilterToSetTests() {
     var intSet = emptyMultiSet<Int>()
     var intExpected = emptyMultiSet<Int>()
